Split tests into separate files.
Added q function again due to it being used elsewhere.
